% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-params.R
\name{generic-params}
\alias{generic-params}
\title{Generic params common to all functions}
\arguments{
\item{.data}{A data frame, data frame extension (e.g., a tibble), or a lazy
data frame (e.g. from dbplyr or dtplyr).}

\item{.nest_data}{A list-column containing data frames}
}
\description{
Generic params common to all functions
}
