% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{job_survey}
\alias{job_survey}
\title{Example survey data regarding job satisfaction}
\format{
A data frame with 500 rows and 6 variables:
\describe{
\item{survey_name}{name of survey}
\item{Q1}{respondent age}
\item{Q2}{city the respondent resides in}
\item{Q3}{field that the respondent that works in}
\item{Q4}{respondent's job satisfaction (on a scale from extremely satisfied to extremely dissatisfied)}
\item{Q5}{respondent's annual salary, in thousands of dollars}
}
}
\usage{
job_survey
}
\description{
A toy dataset containing 500 responses to a job satisfaction survey. The
responses were randomly generated using the Qualtrics survey platform.
}
\keyword{datasets}
