% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_distinct.R
\name{nest_distinct}
\alias{nest_distinct}
\title{Subset distinct/unique rows within a nested tibble or dataframe}
\usage{
nest_distinct(.data, .nest_data, ..., .keep_all = FALSE)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy
lazy data frame (e.g. from dbplyr or dtplyr).}

\item{.nest_data}{A list-col containing a data frames (or tibbles).}

\item{...}{Data masking optional variables to use when determining uniqueness.
if there are multiple rows for a given combination of inputs, only the first
row will be preserved. If omitted, will use all variables.}

\item{.keep_all}{If \code{TRUE}, keep all variables in each nested data frame passed
to \code{.nest_data}. If a combination of \code{...} is not distinct, this keeps the
first row of values.}
}
\value{
A tibble or dataframe.
}
\description{
Select only unique/distinct rows from a nested data frame.
}
\examples{
\dontrun{
library(tidyverse)
library(gapminder)

# return a distinct list of countries within each continent
gapminder \%>\%
  nest(country_data = -continent) \%>\%
  nest_distinct(country_data, country)
}
}
