% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plyr.R
\name{nplyr}
\alias{nplyr}
\alias{.iname}
\alias{.inum}
\alias{.inames}
\alias{.ilength}
\alias{nl_ply}
\alias{nldply}
\alias{nllply}
\alias{na_ply}
\alias{nadply}
\alias{nalply}
\title{Apply Functions over Named Objects}
\usage{
.iname()

.inum()

.inames()

.ilength()

nl_ply(.data, .fun, ...)

nldply(.data, .fun, ...)

nllply(.data, .fun, ...)

na_ply(.data, .margins, .fun, ...)

nadply(.data, .margins, .fun, ...)

nalply(.data, .margins, .fun, ...)
}
\arguments{
\item{.data}{Data object passed to the corresponding \pkg{plyr} function}

\item{.fun}{Function to apply to each -- margin -- element of \code{.data}}

\item{...}{extra arguments passed to corresponding \pkg{plyr} function}

\item{.margins}{Margin over which to apply}
}
\description{
Slightly modified versions of \pkg{plyr}'s functions, which give access
to the current item's name within the applied function.
}
