% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{personal_survey}
\alias{personal_survey}
\title{Example survey data regarding personal life satisfaction}
\format{
A data frame with 750 rows and 6 variables
\describe{
\item{survey_name}{name of survey}
\item{Q1}{respondent age}
\item{Q2}{city the respondent resides in}
\item{Q3}{field that the respondent that works in}
\item{Q4}{respondent's personal life satisfaction (on a scale from extremely satisfied to extremely dissatisfied)}
\item{Q5}{open text response elaborating on personal life satisfaction}
}
}
\usage{
personal_survey
}
\description{
A toy dataset containing 750 responses to a personal satisfaction survey. The
responses were randomly generated using the Qualtrics survey platform.
}
\keyword{datasets}
